Moving storage iterator into core modules.
Fixes #90.
