+++ gcloud/storage/bucket.py
@@ -23,12 +23,12 @@ import six
 
 from gcloud._helpers import get_default_project
 from gcloud.exceptions import NotFound
+from gcloud.iterator import Iterator
 from gcloud.storage._helpers import _PropertyMixin
 from gcloud.storage._helpers import _require_connection
 from gcloud.storage._helpers import _scalar_property
 from gcloud.storage.acl import BucketACL
 from gcloud.storage.acl import DefaultObjectACL
-from gcloud.storage.iterator import Iterator
 from gcloud.storage.blob import Blob
 from gcloud._helpers import _RFC3339_MICROS
